+++ trading-service/src/main/java/com/sajo/trading_service/trading/service/command/OrderExecutionCommandService.java
@@ -10,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Service
@@ -24,7 +25,7 @@ public class OrderExecutionCommandService {
             UUID orderId,
             int totalFilledQuantity,
             int remainingQuantity,
-            long averageExecutionPrice,
+            BigDecimal averageExecutionPrice,
             long totalExecutionAmount
     ) {
         Order order =
@@ -80,7 +81,7 @@ public class OrderExecutionCommandService {
             UUID orderId,
             int totalFilledQuantity,
             int remainingQuantity,
-            long averageExecutionPrice,
+            BigDecimal averageExecutionPrice,
             long totalExecutionAmount
     ) {
         Order order =
